--- src/main/kotlin/com/example/workflow/common/util/ProblemDetailAccessDeniedHandler.kt
@@ -1,25 +1,0 @@
-package com.example.workflow.common.util
-
-import jakarta.servlet.http.HttpServletRequest
-import jakarta.servlet.http.HttpServletResponse
-import org.springframework.http.HttpStatus
-import org.springframework.security.access.AccessDeniedException
-import org.springframework.security.web.access.AccessDeniedHandler
-import org.springframework.stereotype.Component
-
-@Component
-@Suppress("unused")
-class ProblemDetailAccessDeniedHandler : AccessDeniedHandler {
-    override fun handle(
-        request: HttpServletRequest,
-        response: HttpServletResponse,
-        accessDeniedException: AccessDeniedException
-    ) {
-        ProblemDetailWriter.write(
-            request = request,
-            response = response,
-            status = HttpStatus.FORBIDDEN,
-            detail = accessDeniedException.message ?: "Access denied.",
-        )
-    }
-}
