+++ src/main/kotlin/com/example/workflow/infra/security/web/ProblemDetailAccessDeniedHandler.kt
@@ -1,26 +1,26 @@
-package com.example.workflow.common.exception
+package com.example.workflow.infra.security.web
 
 import com.example.workflow.common.util.ProblemDetailWriter
 import jakarta.servlet.http.HttpServletRequest
 import jakarta.servlet.http.HttpServletResponse
 import org.springframework.http.HttpStatus
-import org.springframework.security.core.AuthenticationException
-import org.springframework.security.web.AuthenticationEntryPoint
+import org.springframework.security.access.AccessDeniedException
+import org.springframework.security.web.access.AccessDeniedHandler
 import org.springframework.stereotype.Component
 
 @Component
 @Suppress("unused")
-class ProblemDetailAuthenticationEntryPoint : AuthenticationEntryPoint {
-    override fun commence(
+class ProblemDetailAccessDeniedHandler : AccessDeniedHandler {
+    override fun handle(
         request: HttpServletRequest,
         response: HttpServletResponse,
-        authException: AuthenticationException
+        accessDeniedException: AccessDeniedException
     ) {
         ProblemDetailWriter.write(
             request = request,
             response = response,
-            status = HttpStatus.UNAUTHORIZED,
-            detail = authException.message ?: "Unauthorized error occurred.",
+            status = HttpStatus.FORBIDDEN,
+            detail = accessDeniedException.message ?: "Access denied.",
         )
     }
 }
